requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Scrapy==2.8.0
flask-cors==3.0.10
python-dotenv==1.0.0
requests==2.28.2